Refactor MultiYAxisChart component to improve PDF export functionality
