+++ demo/components/Charts/MultiYAxisChart.tsx
@@ -6,10 +6,11 @@ import jsPDF from "jspdf";
 import { getCurrentDateTime } from "@/utils/dateUtil";
 import { Button } from "primereact/button";
 import { ChartOptions } from "chart.js";
+import { handleExportToPDF } from "@/utils/exportUtil";
 
 const MultiYAxisChart = () => {
     const [chartData, setChartData] = useState({});
-    const [chartOptions, setChartOptions] = useState<ChartOptions<"line"> | undefined>(undefined);
+    const [chartOptions, setChartOptions] = useState<ChartOptions<"line">>({});
     const chartRef = useRef<any>(null);
 
     useEffect(() => {
@@ -154,79 +155,12 @@ const MultiYAxisChart = () => {
         setChartOptions(updateChartOptions("white", "white")); // Initial white theme
     }, []);
 
-    const handleExportToPDF = async () => {
-        const chartInstance = chartRef.current?.getChart();
-        if (chartInstance && chartOptions) {
-            const updatedOptions: ChartOptions<"line"> = {
-                ...chartOptions,
-                plugins: {
-                    ...chartOptions.plugins,
-                    legend: {
-                        labels: { color: "black" },
-                    },
-                },
-                scales: {
-                    x: {
-                        ...chartOptions.scales?.x,
-                        ticks: { color: "black" },
-                        grid: { color: "black" },
-                    },
-                    y: {
-                        ...chartOptions.scales?.y,
-                        ticks: { color: "black" },
-                        grid: { color: "black" },
-                    },
-                    y1: {
-                        ...chartOptions.scales?.y1,
-                        ticks: { color: "black" },
-                        grid: { color: "black", drawOnChartArea: true },                        
-                    },
-                    y2: {
-                        ...chartOptions.scales?.y2,
-                        ticks: { color: "black" },
-                        grid: { color: "black", drawOnChartArea: true },
-                    },
-                    y3: {
-                        ...chartOptions.scales?.y3,
-                        ticks: { color: "black" },
-                        grid: { color:"black", drawOnChartArea: true },
-                    },
-                    y4: {
-                        ...chartOptions.scales?.y4,
-                        ticks: { color: "black" },
-                        grid: { color:"black", drawOnChartArea: true },
-                    },
-                },
-            };
-
-            chartInstance.options = updatedOptions;
-            chartInstance.update();
-
-            await new Promise((resolve) => requestAnimationFrame(resolve));
-
-            const imageData = chartInstance.toBase64Image("image/png", 1.0);
-            const pdf = new jsPDF({
-                orientation: "landscape",
-                unit: "mm",
-                format: "a4",
-            });
-
-            pdf.addImage(imageData, "PNG", 7, 5, 285, 200); // Cover the rotated landscape A4 page 
-            // (10, 40, 297, 210)  is the size of the image on the PDF
-            // x axis, y axis, width, height
-            pdf.save(`${getCurrentDateTime()}.pdf`);
-
-            chartInstance.options = chartOptions;
-            chartInstance.update();
-        } else {
-            alert("Chart instance or options not found.");
-        }
-    };
+    
 
     return (
         <div className="card">
             <Chart ref={chartRef} type="line" data={chartData} options={chartOptions} />
-            <Button label="Export to PDF" onClick={handleExportToPDF} className="p-button-info mt-5" />
+            <Button label="Export to PDF" onClick={() => handleExportToPDF(chartRef, chartOptions)} className="p-button-info mt-5" />
         </div>
     );
 };
